// src/types/auth.ts
export interface TutorSignUpPayload {
  fullName: string;
  phone: string;
  email: string;
  password: string;
}

export interface PetSignUpPayload {
  name: string;
  species: string;
  birthDate: string;
  color: string;
  sex: string;
  weightKg: string;
  size: string;
  microchip: string;
  breed: string;
  notes: string;
}